requirements.txt
streamlit
   Pillow
   google-generativeai
   python-dotenv
   pdfplumber